A bit change in UI, navigation between them done
