+++ app/src/main/AndroidManifest.xml
@@ -12,13 +12,7 @@
         android:theme="@style/AppTheme"
         tools:targetApi="31">
         <activity
-            android:name=".view.activity.ForgotPasswordActivity"
-            android:exported="false" />
-        <activity
             android:name=".view.activity.DemoActivity"
-            android:exported="false" />
-        <activity
-            android:name=".view.activity.LoginActivity"
             android:exported="false" />
         <activity
             android:name=".view.activity.StartupActivity"
@@ -29,8 +23,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ForgotPassword"/>
-        <activity android:name=".Login"/>
+        <activity android:name=".view.activity.ForgotPasswordActivity"/>
+        <activity android:name=".view.activity.LoginActivity"/>
     </application>
 
 </manifest>
