+++ app/src/main/java/com/ubpis/inventame/view/activity/ForgotPasswordActivity.java
@@ -1,8 +1,11 @@
 package com.ubpis.inventame.view.activity;
 
-import androidx.appcompat.app.AppCompatActivity;
-
+import android.content.Intent;
 import android.os.Bundle;
+import android.view.View;
+import android.widget.Toast;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.ubpis.inventame.R;
 
@@ -12,5 +15,17 @@ public class ForgotPasswordActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_forgot_password);
+    }
+
+    public void returnBack(View view){
+        Intent intent = new Intent(ForgotPasswordActivity.this, LoginActivity.class);
+        startActivity(intent);
+    }
+
+    public void send(View view){
+        // si es correcto el correo electronico
+        Toast.makeText(getApplicationContext(), "Se ha enviado el correo correctamente.", Toast.LENGTH_SHORT).show();
+        Intent intent = new Intent(ForgotPasswordActivity.this, LoginActivity.class);
+        startActivity(intent);
     }
 }
