+++ app/src/main/java/com/ubpis/inventame/view/activity/LoginActivity.java
@@ -1,8 +1,10 @@
 package com.ubpis.inventame.view.activity;
 
-import androidx.appcompat.app.AppCompatActivity;
-
+import android.content.Intent;
 import android.os.Bundle;
+import android.view.View;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.ubpis.inventame.R;
 
@@ -13,4 +15,21 @@ public class LoginActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_login);
     }
+
+    public void returnStartUp (View view){
+        Intent intent = new Intent(LoginActivity.this, StartupActivity.class);
+        startActivity(intent);
+    }
+
+    public void recoverPassword (View view){
+        Intent intent = new Intent(LoginActivity.this, ForgotPasswordActivity.class);
+        startActivity(intent);
+    }
+
+    public void loginInventame (View view){
+        //Intent intent = new Intent();
+        //startActivity(intent);
+    }
+
+
 }
