+++ app/src/main/java/com/ubpis/inventame/view/activity/StartupActivity.java
@@ -2,7 +2,9 @@ package com.ubpis.inventame.view.activity;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import android.content.Intent;
 import android.os.Bundle;
+import android.view.View;
 
 import com.ubpis.inventame.R;
 
@@ -12,5 +14,15 @@ public class StartupActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_startup);
+    }
+
+    public void login(View view){
+        Intent intent = new Intent(StartupActivity.this, LoginActivity.class);
+        startActivity(intent);
+    }
+
+    public void register(View view){
+        //Intent intent = new Intent();
+        //startActivity(intent);
     }
 }
